fix a float->enum conversion error with g++ 4
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5748 d7cf8633-e32d-0410-b094-e92efae38249

+++ mvm_cmds.c
@@ -480,7 +480,7 @@ void VM_M_setserverlistmaskstring( void )
 	}
 
 	mask->active = true;
-	mask->tests[field] = (serverlist_maskop_t) PRVM_G_FLOAT( OFS_PARM3 );
+	mask->tests[field] = (serverlist_maskop_t)((int)PRVM_G_FLOAT( OFS_PARM3 ));
 }
 
 /*
@@ -533,7 +533,7 @@ void VM_M_setserverlistmasknumber( void )
 	}
 
 	mask->active = true;
-	mask->tests[field] = (serverlist_maskop_t) PRVM_G_FLOAT( OFS_PARM3 );
+	mask->tests[field] = (serverlist_maskop_t)((int)PRVM_G_FLOAT( OFS_PARM3 ));
 }
